feat: Front Master

See merge request s09-webmobile1-sub2/S09P12E201!137

Former-commit-id: 326ae2ed83fb13deae7ea42c74e1166e220a0443

+++ backend/src/main/java/com/ssafy/api/service/MyBadgeRepository.java
@@ -8,10 +8,10 @@ import org.springframework.stereotype.Repository;
 import com.ssafy.api.model.MyBadge;
 
 @Repository
-public interface MyBadgeRepository extends JpaRepository<MyBadge, Integer> {
+public interface MyBadgeRepository extends JpaRepository<MyBadge, Long> {
 	
-	void deleteByUserID(int userID);
+	void deleteByUserID(long userID);
 
-	List<MyBadge> findAllByUserID(int userID);
+	List<MyBadge> findAllByUserID(long userID);
 
 }
